dapp-update: rearranged contract-interactions component folder
- migrated ExchangeRate and ERC UI components to /component/contract-interaction

- moved deployment support of the ExhcnageRate and ERC into its /deployment
