+++ system-contract-dapp-playground/src/components/contract-interaction/exchange-rate-hip-475/deployment/ExchangeRateDeployField.tsx
@@ -20,9 +20,9 @@
 
 import Image from 'next/image';
 import { useToast } from '@chakra-ui/react';
-import { CommonErrorToast } from '../toast/CommonToast';
 import { Dispatch, SetStateAction, useState } from 'react';
-import HederaCommonTextField from '../common/HederaCommonTextField';
+import { CommonErrorToast } from '../../../toast/CommonToast';
+import HederaCommonTextField from '../../../common/HederaCommonTextField';
 
 interface PageProps {
   isDeploying: boolean;
@@ -30,65 +30,49 @@ interface PageProps {
   setDidDeployStart: Dispatch<SetStateAction<boolean>>;
 }
 
-const ERC20DeployField = ({ isDeploying, setDidDeployStart, setDeployedParams }: PageProps) => {
+const ExchangeRateDeployField = ({
+  isDeploying,
+  setDidDeployStart,
+  setDeployedParams,
+}: PageProps) => {
   const toaster = useToast();
-  const [name, setName] = useState('');
-  const [symbol, setSymbol] = useState('');
+  const [value, setValue] = useState('');
 
   const handleDeploy = async () => {
-    // stop user when the params are empty
-    let description = null;
-    if (name === '' && symbol === '') {
-      description = `Param fields cannot be empty`;
-    } else if (name === '') {
-      description = `Name field cannot be empty`;
-    } else if (symbol === '') {
-      description = `Symbol field cannot be empty`;
-    }
-    if (description) {
+    // stop user when the param is empty
+    if (value === '') {
       CommonErrorToast({
         toaster,
         title: 'Cannot deploy contract',
-        description,
+        description: 'Toll fee cannot be empty',
       });
       return;
     }
 
     // update global deployedParam
-    setDeployedParams([name, symbol]);
+    setDeployedParams([value]);
 
     // trigger deploySmartContract API
     setDidDeployStart(true);
   };
 
   return (
-    <div className="flex flex-col gap-3 items-center justify-center">
+    <div className="flex gap-6 items-center">
       <HederaCommonTextField
         size="md"
-        type="text"
-        title="Name"
-        value={name}
-        setValue={setName}
-        titleBoxSize="w-[200px]"
-        placeholder="e.g. Hedera"
-        explanation="The name of the token"
-      />
-      <HederaCommonTextField
-        size="md"
-        type="text"
-        title="Symbol"
-        value={symbol}
-        setValue={setSymbol}
-        titleBoxSize="w-[200px]"
-        placeholder="e.g. HBAR"
-        explanation="The symbol of the token"
+        title="Toll"
+        value={value}
+        type="number"
+        setValue={setValue}
+        placeholder="e.g. 100"
+        explanation="The USD in cents that must be sent as msg.value"
       />
 
       {/* deploy button */}
       <button
         onClick={handleDeploy}
         disabled={isDeploying}
-        className={`border border-hedera-green text-hedera-green mt-3 px-24 py-2 rounded-xl font-medium hover:bg-hedera-green/50 hover:text-white transition duration-300 ${
+        className={`border border-hedera-green text-hedera-green px-6 py-2 rounded-xl font-medium hover:bg-hedera-green/50 hover:text-white transition duration-300 ${
           isDeploying && `cursor-not-allowed`
         }`}
       >
@@ -111,4 +95,4 @@ const ERC20DeployField = ({ isDeploying, setDidDeployStart, setDeployedParams }:
   );
 };
 
-export default ERC20DeployField;
+export default ExchangeRateDeployField;
